requirements.txt
aiohttp
rich
tldextract
python-whois
